requirements.txt
aiosqlite==0.22.1
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
certifi==2026.4.22
distro==1.9.0
docstring_parser==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
-e git+ssh://git@github.com/gnaniscode/jobcopilot.git@09e1b234e7d4fc407044ce3f4a5d8dc6f06fd8bb#egg=jobcopilot
lxml==6.1.0
pydantic==2.13.3
pydantic_core==2.46.3
pypdf==6.10.2
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
